Refactor: Storing volume as int to avoid doing annoying and pointless byte conversions

+++ src/main/kotlin/model/Instrument.kt
@@ -3,8 +3,8 @@ package model
 data class Instrument(
     val name: String,
     val length: Short,
-    val fineTune: Byte,
-    val volume: Byte,
+    val fineTune: Int,
+    val volume: Int,
     val repeatOffsetStart: Short,
     val repeatLength: Short,
     var floatAudioData: ArrayList<Float> = arrayListOf()
